backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
chromadb>=0.5.0
litellm>=1.50.0
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.50.0
anthropic>=0.39.0
tweepy>=4.14.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
greenlet>=3.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
chardet>=5.0.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
mcp-use>=0.5.0
